+++ moe/easy_interface/__init__.py
@@ -1,2 +1,8 @@
 # -*- coding: utf-8 -*-
-"""An easy, simple interface for using the MOE REST service."""
+"""An easy, simple interface for using the MOE REST service.
+
+Contains:
+    * experiment.py - The a class for historic information required to optimize using MOE
+    * simple_endpoint.py - An example use of the REST interface using python
+
+"""
