+++ moe/views/gp_next_points_constant_liar.py
@@ -40,7 +40,7 @@ class GpNextPointsconstant_liar(GpNextPointsPrettyView):
 
     @view_config(route_name=pretty_route_name, renderer=GpNextPointsPrettyView.pretty_renderer)
     def pretty_view(self):
-        """A pretty, browser interactable view for the interface. Includes form request and response."""
+        """A pretty, browser interactive view for the interface. Includes form request and response."""
         return self.pretty_response()
 
     @view_config(route_name=route_name, renderer='json', request_method='POST')
